requirements.txt
pandas
geopandas
matplotlib
requests
numpy
scipy
rasterio
rasterstats
python-dotenv
